Add CriteriaPage and view/update meetings

+++ frontend/src/pages/MeetingsPage.tsx
@@ -1,31 +1,19 @@
 import React, {useEffect, useState} from 'react';
-import axios from 'axios';
 import {useNavigate} from 'react-router-dom';
 import {getMeetings, deleteMeeting} from '../services/apiService';
 import {format} from 'date-fns';
 import {ru} from 'date-fns/locale';
 
-const formatDate = (date) => {
-    return format(new Date(date), 'dd MMMM yyyy, HH:mm', {locale: ru});
+export const formatDate = (date) => {
+    return format(new Date(date), 'd MMMM yyyy, HH:mm', {locale: ru});
 }
 
 export function MeetingsPage() {
     const [meetings, setMeetings] = useState([]);
     const navigate = useNavigate();
 
-    const fetchMeetings = async () => {
-        const token = sessionStorage.getItem('token');
-
-        if (!token) {
-            navigate("/login", {replace: true});
-            return;
-        }
-
-        getMeetings().then(response => setMeetings(response.data));
-    }
-    
     useEffect(() => {
-        fetchMeetings();
+        getMeetings().then(response => setMeetings(response.data));
     }, []);
 
     const handleNewMeeting = () => {
@@ -33,8 +21,15 @@ export function MeetingsPage() {
     }
 
     const handleDeleteMeeting = async (id) => {
-        await deleteMeeting(id);
-        fetchMeetings();
+        const isConfirmed = window.confirm('Вы уверены, что хотите удалить это заседание?');
+        if (isConfirmed) {
+            await deleteMeeting(id);
+            getMeetings().then(response => setMeetings(response.data));
+        }
+    }
+
+    const handleViewMeeting = async (id) => {
+        navigate(`/meetings/${id}`, {replace: true});
     }
 
     return (
@@ -52,12 +47,17 @@ export function MeetingsPage() {
                     <div key={meeting.id} className="col-md-4 mb-4">
                         <div className="card">
                             <div className="card-body">
-                                <h5 className="card-title">{meeting.info}</h5>
+                                <h4 className="card-title pb-2 text-primary">{meeting.info}</h4>
                                 <p className="card-text">
                                     <strong>Дата:</strong> {formatDate(meeting.dateAndTime)}<br/>
                                     <strong>Аудитория:</strong> {meeting.auditorium}
                                 </p>
-                                <button className="btn btn-outline-danger" onClick={() => handleDeleteMeeting(meeting.id)}>
+                                <button className="btn btn-outline-primary me-2"
+                                        onClick={() => handleViewMeeting(meeting.id)}>
+                                    Перейти
+                                </button>
+                                <button className="btn btn-outline-danger"
+                                        onClick={() => handleDeleteMeeting(meeting.id)}>
                                     Удалить
                                 </button>
                             </div>
